// frontend/src/client/api/generated/model/enumCategoryType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Marumie Backend
 * OpenAPI spec version: 0.1.0
 */

export type EnumCategoryType =
  (typeof EnumCategoryType)[keyof typeof EnumCategoryType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EnumCategoryType = {
  income: "income",
  expense: "expense",
} as const;